+++ src/App.jsx
@@ -2,6 +2,7 @@ import TodoCreate from 'components/TodoCreate';
 import TodoHeader from 'components/TodoHeader';
 import TodoList from 'components/TodoList';
 import { useEffect, useReducer } from 'react';
+import todoReducer from 'reducer/todoReducer';
 import styled from 'styled-components';
 import { flexCenter, flexColumn } from 'styles/commonStyle';
 
@@ -10,49 +11,12 @@ const initialState = {
   doneList: JSON.parse(localStorage.getItem('doneList')) || [],
 };
 
-const todoReducer = (state, action) => {
-  switch (action.type) {
-    case 'ADD_TODO':
-      return { ...state, todoList: [...state.todoList, action.payload] };
-    case 'MOVE_TODO_TO_DONE':
-      return moveItem(state, action.payload, 'todoList', 'doneList');
-    case 'MOVE_DONE_TO_TODO':
-      return moveItem(state, action.payload, 'doneList', 'todoList');
-    case 'REMOVE_TODO':
-      return removeItem(state, action.payload, 'todoList');
-    case 'REMOVE_DONE':
-      return removeItem(state, action.payload, 'doneList');
-    default:
-      return state;
-  }
-};
-
-const moveItem = (state, payload, fromListName, toListName) => {
-  const { id } = payload;
-
-  return {
-    [fromListName]: state[fromListName].filter((todo) => todo.id !== id),
-    [toListName]: [...state[toListName], payload],
-  };
-};
-
-const removeItem = (state, payload, listName) => {
-  const { id } = payload;
-  return {
-    ...state,
-    [listName]: state[listName].filter((todo) => todo.id !== id),
-  };
-};
-
 function App() {
   const [state, dispatch] = useReducer(todoReducer, initialState);
   const { todoList, doneList } = state;
 
   useEffect(() => {
     localStorage.setItem('todoList', JSON.stringify(todoList));
-  }, [todoList]);
-
-  useEffect(() => {
     localStorage.setItem('doneList', JSON.stringify(doneList));
   }, [todoList]);
